backend/pyproject.toml
[project]
name = "jaram-paper-backend"
version = "1.0.0"
description = "Jaram Paper Backend API"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "sqlalchemy==2.0.36",
    "feedparser==6.0.11",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "pydantic==2.9.2",
    "arxiv==2.1.3",
    "openai==1.54.0",
    "schedule==1.2.2",
    "python-multipart==0.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
